Up coverage + add jacoco-config.gradle + refactor names
