+++ src/main/java/ua/marketplace/security/CustomUserDetailsService.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import ua.marketplace.entitys.User;
+import ua.marketplace.entities.User;
 import ua.marketplace.repositoryes.UserRepository;
 
 import java.util.Collections;
