+++ src/main/java/ua/marketplace/services/PhoneNumberRegistrationService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 import ua.marketplace.data.Error;
 import ua.marketplace.dto.CodeDto;
 import ua.marketplace.dto.PhoneNumberDto;
-import ua.marketplace.entitys.User;
+import ua.marketplace.entities.User;
 import ua.marketplace.repositoryes.UserRepository;
 import ua.marketplace.requests.PhoneCodeRequest;
 import ua.marketplace.requests.PhoneNumberRequest;
