+++ src/test/java/ua/marketplace/controllers/RegistrationTest.java
@@ -17,7 +17,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.validation.MapBindingResult;
 import ua.marketplace.dto.CodeDto;
 import ua.marketplace.dto.PhoneNumberDto;
-import ua.marketplace.entitys.User;
+import ua.marketplace.entities.User;
 import ua.marketplace.repositoryes.UserRepository;
 import ua.marketplace.requests.PhoneCodeRequest;
 import ua.marketplace.requests.PhoneNumberRequest;
